Allow * in CORS
CORS * allows any web client to send requests to the API.
The API will check for the token so:
- browsers sending an authorization token will get access to protected
  content;
- browsers not sending an authorization token will only get access to
  public content.

There is no security reason to maintain and restrict access to some
whitelisted domains.

When using cookies together with a CSRF token, the situation is
identical and using * is also secure.