pull in gunicorn
The idea is to avoid using a different WSGI server between dev & prod,
and gunicorn has a few production-related features which lack from
waitress. Also, I'd like to drop apache in favor of haproxy.

First step: evaluate if gunicorn offers a valid alternative for both
use-cases by running
  `.build/venv/bin/gunicorn --paste development.ini -b :6543`
instead of
  `make serve`

Once we've validated the use of this tool, we can update the
`[server:main]` option and get rid of the legacy bits (waitress, apache,
etc).