Only refresh requirements.txt when needed
Currently, each time we run the Makefile with the serve target,
the egg target is run and the requirements.txt are refreshed.
This is due to the egg-link never being created.

Remove broken egg target.
Depend on pip file instead of a directory.
Only trigger a requirements.txt refresh when needed.

This commit is equivalent to what was done for the UI.