Remove duplicates in recent outings
As usual, I do not know python and the pyramid framework, nor am I testing this.
But we know the request can return duplicates (see e.g. Meij waypoint with says it has >2000 outings, but really only ~800), I assume this could fix this.
As a consequence we could enhance UI a little bit and remove some useless code on the frontend part.