Rework of development environment and docker image build:
See HOWTO-DEV.md to understand how the dev environment should now be setup and run.
The changes eliminate py3compile, the need for a full compilation env in the docker image and on the developer's machine.
It keeps the Makefile as a shortcut to run some tasks in dev, but does not make it mandatory.